Verilog: extract top_level_modules into separate file
This moves the discovery of top-level Verilog modules into a separate
function and separate fule.
