+++ regression/verilog/comments/eof-in-comment1.desc
@@ -1,7 +1,7 @@
 CORE
 eof-in-comment1.v
 
-^no module found$
+^error: no module found$
 ^EXIT=1$
 ^SIGNAL=0$
 --
