+++ regression/verilog/constraints/constraint1.desc
@@ -1,7 +1,7 @@
 CORE
 constraint1.sv
 
-^no module found$
+^error: no module found$
 ^EXIT=1$
 ^SIGNAL=0$
 --
