+++ src/verilog/Makefile
@@ -2,6 +2,7 @@ SRC = aval_bval_encoding.cpp \
       convert_literals.cpp \
       expr2verilog.cpp \
       sva_expr.cpp \
+      top_level_modules.cpp \
       typename.cpp \
       verilog_bits.cpp \
       verilog_ebmc_language.cpp \
